Add manual for Lego Star Wars: The Complete Saga
This is a repack of Roush's v7.1 Lego Star Wars: The Complete Saga
manual world:
https://discord.com/channels/1097532591650910289/1156373916370083880/1290558889963552789

The repack uses a work-in-progress fork of manual that improves logic
performance. Because this fork is based on neither a stable nor unstable
release of manual, the included manual client is given the lowest
possible priority, causing manual clients from other loaded manual
worlds to be loaded in preference to the client included in this world.